Refactor ecma signature encoding logic out of ManagedBinaryEmitter - Create signature encoder that is useable with multiple sources of valid tokens
[tfs-changeset: 1676780]
